+++ engine/src/main/java/com/mirafintech/prototype/model/charge/TransactionFee.java
@@ -16,7 +16,7 @@ import java.time.LocalDateTime;
 @Getter
 public final class TransactionFee extends MerchantCharge {
 
-    @Column(name = "transaction_fee__amount", precision = 13, scale = 5)
+    @Column(name = "transaction_fee__amount", precision = 16, scale = 5)
     private BigDecimal amount;
 
     protected TransactionFee() {
